优化友盟 SDK 的登录处理逻辑

+++ library/umengSdk/src/main/java/com/hjq/umeng/sdk/UmengClient.kt
@@ -97,14 +97,15 @@ object UmengClient {
         }
 
         try {
-            // 删除旧的第三方登录授权
-            UMShareAPI.get(activity).deleteOauth(activity, platform.getThirdParty(), null)
-            // 要先等上面的代码执行完毕之后
-            Thread.sleep(200)
+            if (UMShareAPI.get(activity).isAuthorize(activity, platform.getThirdParty())) {
+                // 删除旧的第三方登录授权
+                UMShareAPI.get(activity).deleteOauth(activity, platform.getThirdParty(), null)
+                // 要先等上面的代码执行完毕之后
+                Thread.sleep(200)
+            }
             // 开启新的第三方登录授权
             UMShareAPI.get(activity).getPlatformInfo(activity, platform.getThirdParty(),
-                UmengLogin.LoginListenerWrapper(requireNotNull(platform.getThirdParty()), listener)
-            )
+                UmengLogin.LoginListenerWrapper(requireNotNull(platform.getThirdParty()), listener))
         } catch (e: InterruptedException) {
             e.printStackTrace()
         }
